resize: Port to the new upstreamed APIs and add options
Most of the old code has been upstreamed. I used the reduced
complexity to add tons of new visual options. It looks great!

This remove support for resizing tiled layouts. First of all, it
never worked properly. Also, I have a new layout system with
a saner way of doing this. As it wont make it for 3.6, I will
merge it into Collision.