Fix bug in focus module when no client is currently focused
- Suppose we have 2 screens side-by-side and the mouse is at the
screen 1 (left) and no client is open/focused there, and the user
requests to move focus to the client at right (screen 2). Currently
the focus is moved correctly, but the mouse stays in the screen one,
leading to kinda inconsistence for the next commands, etc.
- This patch modifies the focus.bydirection() to move mouse to new
client's screen (we're moving focus to) even when there was no client
focused at that point.

Signed-off-by: Nick Diego Yamane <nick@diegoyam.com>