Fixes errors in max layout - part 2
If one of the clients was maximised (like PDF viewer) an error was
thrown and visual artifact left on the screen.

This is the second patch that changes awful.client.tiled to
awful.client.visible.

Without this change a preview is displayed only for clients in tiled mode
but the navigation (left, right) is done on ALL clients (including
clients in full screen).